refactor: move more internal fns into _internal folders
