+++ packages/common/src/sites/_internal/_purge-non-guids-from-catalog.ts
@@ -1,4 +1,4 @@
-import { IModel, getProp, cloneObject, isGuid } from "..";
+import { IModel, getProp, cloneObject, isGuid } from "../..";
 
 /**
  * Remove any non-guid entries from the data catalog groups array
