+++ packages/common/src/sites/upgrade-site-schema.ts
@@ -1,12 +1,12 @@
 import { IModel } from "../types";
 import { getProp } from "../objects";
 import { SITE_SCHEMA_VERSION } from "./site-schema-version";
-import { _applySiteSchema } from "./_apply-site-schema";
-import { _enforceLowercaseDomains } from "./_enforce-lowercase-domains";
-import { _ensureCatalog } from "./_ensure-catalog";
-import { _purgeNonGuidsFromCatalog } from "./_purge-non-guids-from-catalog";
-import { _ensureTelemetry } from "./_ensure-telemetry";
-import { _migrateFeedConfig } from "./_migrate-feed-config";
+import { _applySiteSchema } from "./_internal/_apply-site-schema";
+import { _enforceLowercaseDomains } from "./_internal/_enforce-lowercase-domains";
+import { _ensureCatalog } from "./_internal/_ensure-catalog";
+import { _purgeNonGuidsFromCatalog } from "./_internal/_purge-non-guids-from-catalog";
+import { _ensureTelemetry } from "./_internal/_ensure-telemetry";
+import { _migrateFeedConfig } from "./_internal/_migrate-feed-config";
 
 /**
  * Upgrades the schema upgrades
