+++ packages/common/test/sites/_internal/_ensure-telemetry.test.ts
@@ -6,11 +6,11 @@ import {
   mergeObjects,
   setProp,
   buildDraft,
-} from "../../src";
+} from "../../../src";
 import {
   draftModelOneThree,
   oneThreeSiteDraftIncludeList,
-} from "../fixtures/historical-site-draft-schemas/1-3";
+} from "../../fixtures/historical-site-draft-schemas/1-3";
 
 const siteModel = {
   item: {
