+++ packages/common/test/sites/upgrade-site-schema.test.ts
@@ -1,10 +1,10 @@
 import { upgradeSiteSchema } from "../../src";
-import * as _applySiteSchemaModule from "../../src/sites/_apply-site-schema";
-import * as _enforceLowercaseDomainsModule from "../../src/sites/_enforce-lowercase-domains";
-import * as _ensureCatalogModule from "../../src/sites/_ensure-catalog";
-import * as _purgeNonGuidsFromCatalogModule from "../../src/sites/_purge-non-guids-from-catalog";
-import * as _ensureTelemetryModule from "../../src/sites/_ensure-telemetry";
-import * as _migrateFeedConfigModule from "../../src/sites/_migrate-feed-config";
+import * as _applySiteSchemaModule from "../../src/sites/_internal/_apply-site-schema";
+import * as _enforceLowercaseDomainsModule from "../../src/sites/_internal/_enforce-lowercase-domains";
+import * as _ensureCatalogModule from "../../src/sites/_internal/_ensure-catalog";
+import * as _purgeNonGuidsFromCatalogModule from "../../src/sites/_internal/_purge-non-guids-from-catalog";
+import * as _ensureTelemetryModule from "../../src/sites/_internal/_ensure-telemetry";
+import * as _migrateFeedConfigModule from "../../src/sites/_internal/_migrate-feed-config";
 import { IModel } from "../../src";
 import { SITE_SCHEMA_VERSION } from "../../src/sites/site-schema-version";
 import { expectAllCalled, expectAll } from "./test-helpers.test";
